Sphere_planner/backend_python/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymongo==4.16.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==1.3.1
